requirements.txt
transformers~=4.41.2
git+https://github.com/openai/CLIP.git
pillow
termcolor
flatbuffers
#for deployment
fastapi
uvicorn